campaign_intelligence_v2/requirements.txt
# ── FastAPI & server ───────────────────────────────────────────────────────
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-dotenv==1.0.1

# ── LangChain / LangGraph ──────────────────────────────────────────────────
langchain==0.3.7
langchain-core==0.3.15
langchain-openai==0.2.6
langgraph==0.2.45

# ── Pydantic ───────────────────────────────────────────────────────────────
pydantic==2.9.2

# ── OpenAI SDK ─────────────────────────────────────────────────────────────
openai==1.54.4

# ── Export ─────────────────────────────────────────────────────────────────
python-docx==1.1.2
python-pptx==1.0.2

# ── Utilities ──────────────────────────────────────────────────────────────
pandas==2.2.3
